requirements.txt
streamlit
google-genai
chromadb
pypdf
python-dotenv